requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
feedparser>=6.0.11

# Data processing (latest versions that support Python 3.13)
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0

# NLP (using latest versions with potential Python 3.13 support)
spacy>=3.7.0
transformers>=4.36.0
sentence-transformers>=2.2.2

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# Streamlit (latest stable version)
streamlit>=1.29.0
